+++ CHANGELOG.rst
@@ -6,7 +6,8 @@ Changelog
 
 * :issue:`25`: Add new extension ``sphinx-issues`` to Read the Docs.
 * :issue:`26`: Add new extension ``sphinx.ext.linkcode`` to Read the Docs.
-* Add ``method`` argument to :meth:`api.utils.sort_alleles` method. When ``method='name'`` it will sort star alleles by allele number.
+* Add ``by`` argument to :meth:`api.utils.sort_alleles` method. When ``by='name'`` it will sort star alleles by allele number.
+* Update :command:`call-genotypes` command to output genotypes with number-sorted alleles (e.g. '\*4/\*10' instead of '\*10/\*4').
 
 0.5.0 (2021-10-02)
 ------------------